+++ Modules/Intent.Modules.VisualStudio.Projects/Api/Extensions/ServiceFabricType.cs
@@ -1,0 +1,8 @@
+﻿namespace Intent.Modules.VisualStudio.Projects.Api.Extensions;
+
+public enum ServiceFabricType
+{
+    Actor,
+    Stateful,
+    Stateless
+}