+++ Modules/Intent.Modules.VisualStudio.Projects/Templates/ServiceFabric/PublishProfileLocal1Node/PublishProfileLocal1NodeTemplateRegistration.cs
@@ -13,19 +13,19 @@ using Intent.Templates;
 [assembly: DefaultIntentManaged(Mode.Fully)]
 [assembly: IntentTemplate("Intent.ModuleBuilder.TemplateRegistration.Custom", Version = "1.0")]
 
-namespace Intent.Modules.VisualStudio.Projects.Templates.ServiceFabric.PackagesConfig
+namespace Intent.Modules.VisualStudio.Projects.Templates.ServiceFabric.PublishProfileLocal1Node
 {
     [IntentManaged(Mode.Merge, Body = Mode.Merge, Signature = Mode.Fully)]
-    public class PackagesConfigTemplateRegistration : ITemplateRegistration
+    public class PublishProfileLocal1NodeTemplateRegistration : ITemplateRegistration
     {
         private readonly IMetadataManager _metadataManager;
 
-        public PackagesConfigTemplateRegistration(IMetadataManager metadataManager)
+        public PublishProfileLocal1NodeTemplateRegistration(IMetadataManager metadataManager)
         {
             _metadataManager = metadataManager;
         }
 
-        public string TemplateId => PackagesConfigTemplate.TemplateId;
+        public string TemplateId => PublishProfileLocal1NodeTemplate.TemplateId;
 
         [IntentManaged(Mode.Fully, Body = Mode.Ignore)]
         public void DoRegistration(ITemplateInstanceRegistry registry, IApplication applicationManager)
@@ -35,7 +35,7 @@ namespace Intent.Modules.VisualStudio.Projects.Templates.ServiceFabric.PackagesC
             foreach (var model in models)
             {
                 var project = applicationManager.Projects.Single(x => x.Id == model.Id);
-                registry.Register(TemplateId, project, p => new PackagesConfigTemplate(p, model));
+                registry.Register(TemplateId, project, p => new PublishProfileLocal1NodeTemplate(p, model));
             }
         }
     }